update shared/slotutil and shared/testutil

+++ shared/testutil/state.go
@@ -12,7 +12,7 @@ var st, _ = stateTrie.InitializeFromProtoUnsafe(&pb.BeaconState{
 	BlockRoots:                 filledByteSlice2D(params.BeaconConfig().SlotsPerHistoricalRoot, 32),
 	StateRoots:                 filledByteSlice2D(params.BeaconConfig().SlotsPerHistoricalRoot, 32),
 	Slashings:                  make([]uint64, params.BeaconConfig().EpochsPerSlashingsVector),
-	RandaoMixes:                filledByteSlice2D(params.BeaconConfig().EpochsPerHistoricalVector, 32),
+	RandaoMixes:                filledByteSlice2D(uint64(params.BeaconConfig().EpochsPerHistoricalVector), 32),
 	Validators:                 make([]*ethpb.Validator, 0),
 	CurrentJustifiedCheckpoint: &ethpb.Checkpoint{Root: make([]byte, 32)},
 	Eth1Data: &ethpb.Eth1Data{
